gigamart-project/requirements.txt
apache-beam[gcp]>=2.69.0

pyyaml>=6.0

google-cloud-bigquery>=3.11.0
google-cloud-pubsub>=2.18.0
google-cloud-storage>=2.16.0

pytest>=7.4.0
pytest-cov>=4.1.0
